Added dispose(); now caching global; changed global->sandbox reference.
dispose() frees the context memory for a sandbox.

ContextifyInfo now stores a strong reference to the sandbox, which means
it won't be GC'd until dispose() is called.  The flip side is that the
sandbox won't go out of scope if a user drops their sandbox reference,
but keeps a global reference.

We now cache the global object inside of ContextifyInfo instead of
looking it up every time.

Added tests and updated README for dispose().