Fix for #13: Can't use global.eval as a function.
The issue is caused because of a security check in V8's eval
implementation that's only performend when calling eval from an alias:
    if (!this_is_global_receiver || global_is_detached)

If that check fails, it throws the exception from #13.

Previously, we were detaching the proxy global and running the context
on just the real global, but that approach fails the V8 security check
since global_is_detached is true.

To fix it, we no longer detach the proxy global from the context. We
used to do this because setting properties on the proxy global would not
propagate them to the "real" global, which is the object constructed
from our ObjectTemplate with the Named Property Handlers.

It turns out, the proxy global *will* forward everything to the real
global like it's supposed to, but only if you supply/activate
[Named|Indexed]SecurityCallbacks on the ObjectTemplate by using
SetAccessCheckCallbacks.

Adding SecurityCallbacks that always return true allows the proxy global
to forward accesses to the real global, so we don't need to detach the
proxy global, and therefore eval works in all cases.